Repository re-org.
Changing the structure of the repos around. Now CCPP is the only
repository, the IPD is part of CCPP.

All physics schemes are in a sub-directory under `src/schemes`.
This allows for simple schemes to be in a single file while
other large schemes can have thier own sub-directory and the
ability to have their own external build system.

Removed the creation of a shared library for CCPP. At this stage
we are going to just dlopen ourselves. Added a cmake check
for the need to use export dynamic at link time (ELF vs Mach-O).

Added an field index mapper (using `bsearch()`). Tried to simplify
the adding and finding of fields for the caps.