Overhaul of the CCPP.
Decided to rename ALL functions and subroutines to be prefixed
with ccpp_. This also involved renaming all source files to be
prefixed as well.

Updated the fields index routines so as to add a field without
having to specify the rank and shape.

Created a test directory within the src directory to hold all
the test cases.

Rejigged the schema XML parser to be able to handle the
different/optional fields and attributes of the schema.