Adding a ccpp toplevel driver.
Adding a ccpp toplevel driver/init/fini routine for the library.
Removing the C types header as it is not needed anymore.

Expanded the CCPP fields subroutines to include integer 32, 64 and
real 32, 64 bits. Using iso_fortran_env, so as to be able to remove
the kinds module in the future.

Expanded the use statements in the ccpp_ipd module to one pull in
what it needs.

Renamed the ccpp_phy_infra module to be ccpp_suite as that is what
it really focuses on.