Clean-ups and bug-fixes.
Cleaned-up the CMakeLists.txt, so the project installs properly.

Bug fix in ccpp_suite with assumed lenght strings.

Added a logical data field to the ccpp_fields

Cleaned-up the test case, all arrays (in the atm cap) have to have the
target attribute and all arrays (in the physics cap) must have the
pointer attribute.