Clean-up in prep for v0.0.1 release.
A pretty big clean-up, hopefully getting ready for v0.0.1 release.

Since we re-jigged the source tree (moving the schemes out),
we don't need to use the cmake "add_sources()" function to push
source files up to their parent scope. I'll leave the macro/function
in for now.

Renaming the toplevel tests to be examples.

In moving the schemes up to the toplevel, I've renamed the check
one and made it into it's own library.

Going forward, each scheme should be it's own library. This means
the application has to know where the library is so as to open it
(i.e. LD_LIBRARY_PATH or RPATH).

Since scheme should be it's own library, we don't dlopen() ourself
and so there is no need to add the "-Wl,--export-dynamic" flag
to the linker on Linux.

A simple todo would be to make the library (dlopen() call),
use the absolute path if provided.

Adding the optional "lib" and "ver" attributes to the scheme element
in the XSD (and a few XML files).

Added a ccpp_ipd_init() and ccpp_ipd_fini() subroutines.
- The init will dlopen() the library and dlsym() to find the function.
- The fini closes the library.
This is make sure we are able to call the scheme cap function when needed.

Revamped the ccpp_ipd_run() subroutine to ONLY call the ONE scheme it
was asked to call instead of looping over ALL schemes. We are putting
the scheme  calling logic into the hands of the caller.

TODO:
Possible to use c_funptr's instead of generic c_ptr and therefore expose
the cap routines to Fortran...