Modifying the README.md.
Updating the Travis and Coveralls URL (we still need to enable this
for NCAR).

Reformatting to keep paragraphs under 80 columns.

The `$(pwd)` within a code fragment was not rendering, so changed
it to an env var `${PWD}`.

Changed the code blocks to use fenced code block markers for doxygen.

Removed a repeated paragraph on using the `test_init_fini.f90`
program.

Change a reference to IPD to CCPP.