Adding init and fini physics subroutine ability.
A suite can have an init and fini key words in the XML schema.
This meant I re-worked how the XML is parsed and moved the
element parsing into the ccpp_xml.f90 file (out of the
ccpp_suite.f90). This reduces the ugliness of the suite init
subroutine.

The library keyword for the suite XML is now required.

Also changed the STR_LEN to be CCPP_STR_LEN to be consistent
with the CCPP naming convention.