Add support for scheme w/o run phase (#548)
Here we include a small fix to allow for schemes without run phases.

Capgen assumes that all schemes have a run phase, which is not true in practice. During the "scheme analyze" section, if a scheme doesn't contain a run phase we set a flag to true (no_run_phase). Then this flag is queried later in the "scheme write" section.

There are no interface changes.

Addresses #543

Testing has been expanded to contain a scheme w/o a run phase.