Fix for ccpp_track_variables.py: Need "recursive=True" for glob on .meta files (#581)
Because of the change in directory structure for CCPP physics
(https://github.com/ufs-community/ccpp-physics/pull/99), there are now
`.meta` files at different levels in the directory tree. The
`ccpp_track_variables.py` script needs the location of these `.meta`
files as an input argument to the script, but the call to `glob.glob` in
the script does not use the `recursive=True` argument, so even if the
user passes in the argument `-m './physics/physics/**/'` (which should
include all subdirectories), the call to `glob.glob` only searches one
level. Our simple test case only has `.meta` files at a single directory
level, so we never caught this issue.

Simply adding the `recursive=True` argument fixes this issue.