New implementation of differing module name (#646)
New implementation of differing module name

Implement a solution that allows a Fortran module name to differ from the filename.

An optional module_name keyword is added to the ccpp-table-properties (MetadataTable) section that allows the user to specify a Fortran module name that is independent from the name of the enclosing file.